Apply reviewed development package
