+++ .github/scripts/audit_runtime_stress.mjs
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 "use strict";
 
+import crypto from "node:crypto";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
@@ -8,29 +9,31 @@ import { spawnSync } from "node:child_process";
 import { fileURLToPath } from "node:url";
 
 const root = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../..");
-const output = process.env.MCMS_AUDIT_OUTPUT || path.join(root, "audit-output");
-fs.mkdirSync(output, { recursive: true });
+const preflightPath = path.join(root, ".github/scripts/run_userscript_preflight.sh");
+const integrationContract = ".github/scripts/test_ui_mount_integration.mjs";
 
-const defaultRepeats = Number.parseInt(process.env.MCMS_AUDIT_REPEATS || "8", 10);
-const tests = [
-  [".github/scripts/test_transport_sweep_runtime.js", defaultRepeats],
-  [".github/scripts/test_issue523_transport_sweep_progress_runtime.js", defaultRepeats],
-  [".github/scripts/test_issue527_transport_sweep_skipped_patients_runtime.js", defaultRepeats],
-  [".github/scripts/test_issue530_transport_sweep_discharge_confirmation_runtime.js", defaultRepeats],
-  [".github/scripts/test_issue565_transport_sweep_no_reward_runtime.mjs", defaultRepeats],
-  [".github/scripts/test_issue515_launcher_runtime.js", defaultRepeats],
-  [".github/scripts/test_issue517_incident_command_wire_runtime.js", defaultRepeats],
-  [".github/scripts/test_issue553_alliance_member_manager_menu_runtime.js", defaultRepeats],
-  [".github/scripts/test_issue553_alliance_member_manager_page_runtime.js", defaultRepeats],
-  [".github/scripts/test_ui_mount_integration.mjs", Math.max(defaultRepeats, 16)],
-];
+export function discoverRuntimeContracts(preflightSource) {
+  const discovered = [];
+  const seen = new Set();
+  for (const line of String(preflightSource || "").split(/\r?\n/u)) {
+    const match = line.match(/^\s*node\s+(\.github\/scripts\/test_[A-Za-z0-9_./-]+\.(?:js|mjs))(?:\s.*)?$/u);
+    if (!match) continue;
+    const relative = match[1];
+    if (seen.has(relative)) continue;
+    seen.add(relative);
+    discovered.push(relative);
+  }
+  if (!discovered.length) throw new Error("Canonical preflight contains no Node runtime contracts.");
+  return discovered;
+}
 
-const perRunSecondsCeiling = Number.parseFloat(process.env.MCMS_AUDIT_RUN_SECONDS_CEILING || "45");
-const peakRssKbCeiling = Number.parseInt(process.env.MCMS_AUDIT_RSS_KB_CEILING || "700000", 10);
-const totalSecondsCeiling = Number.parseFloat(process.env.MCMS_AUDIT_TOTAL_SECONDS_CEILING || "900");
-const results = [];
-const failures = [];
-const started = process.hrtime.bigint();
+export function buildRuntimeStressPlan(preflightSource, defaultRepeats = 8) {
+  const repeats = Math.max(1, Number.parseInt(String(defaultRepeats), 10) || 8);
+  const contracts = discoverRuntimeContracts(preflightSource);
+  const plan = contracts.map(relative => [relative, repeats]);
+  if (!contracts.includes(integrationContract)) plan.push([integrationContract, Math.max(repeats, 16)]);
+  return plan;
+}
 
 function median(values) {
   if (!values.length) return 0;
@@ -39,112 +42,127 @@ function median(values) {
   return sorted.length % 2 ? sorted[middle] : (sorted[middle - 1] + sorted[middle]) / 2;
 }
 
-for (const [relative, repeats] of tests) {
-  const absolute = path.join(root, relative);
-  if (!fs.existsSync(absolute)) {
-    failures.push(`${relative}: missing runtime contract`);
-    continue;
-  }
+export function runRuntimeStressAudit() {
+  const output = process.env.MCMS_AUDIT_OUTPUT || path.join(root, "audit-output");
+  fs.mkdirSync(output, { recursive: true });
+  const defaultRepeats = Number.parseInt(process.env.MCMS_AUDIT_REPEATS || "8", 10);
+  const preflightSource = fs.readFileSync(preflightPath, "utf8");
+  const tests = buildRuntimeStressPlan(preflightSource, defaultRepeats);
+  const perRunSecondsCeiling = Number.parseFloat(process.env.MCMS_AUDIT_RUN_SECONDS_CEILING || "45");
+  const peakRssKbCeiling = Number.parseInt(process.env.MCMS_AUDIT_RSS_KB_CEILING || "700000", 10);
+  const totalSecondsCeiling = Number.parseFloat(process.env.MCMS_AUDIT_TOTAL_SECONDS_CEILING || "900");
+  const results = [];
+  const failures = [];
+  const started = process.hrtime.bigint();
 
-  const samples = [];
-  for (let iteration = 1; iteration <= repeats; iteration += 1) {
-    const metricsPath = path.join(os.tmpdir(), `mcms-audit-${process.pid}-${samples.length}-${iteration}.txt`);
-    const run = spawnSync(
-      "/usr/bin/time",
-      ["-f", "%e %M", "-o", metricsPath, process.execPath, "--unhandled-rejections=strict", absolute],
-      {
-        cwd: root,
-        encoding: "utf8",
-        timeout: Math.ceil(perRunSecondsCeiling * 1000) + 15000,
-        env: {
-          ...process.env,
-          NODE_OPTIONS: [process.env.NODE_OPTIONS, "--unhandled-rejections=strict"].filter(Boolean).join(" "),
+  for (const [relative, repeats] of tests) {
+    const absolute = path.join(root, relative);
+    if (!fs.existsSync(absolute)) {
+      failures.push(`${relative}: missing runtime contract`);
+      continue;
+    }
+    const samples = [];
+    for (let iteration = 1; iteration <= repeats; iteration += 1) {
+      const metricsPath = path.join(os.tmpdir(), `mcms-audit-${process.pid}-${results.length}-${iteration}.txt`);
+      const run = spawnSync(
+        "/usr/bin/time",
+        ["-f", "%e %M", "-o", metricsPath, process.execPath, "--unhandled-rejections=strict", absolute],
+        {
+          cwd: root,
+          encoding: "utf8",
+          timeout: Math.ceil(perRunSecondsCeiling * 1000) + 15000,
+          env: {
+            ...process.env,
+            NODE_OPTIONS: [process.env.NODE_OPTIONS, "--unhandled-rejections=strict"].filter(Boolean).join(" "),
+          },
+          maxBuffer: 16 * 1024 * 1024,
         },
-        maxBuffer: 16 * 1024 * 1024,
-      },
-    );
-
-    let elapsedSeconds = Number.NaN;
-    let peakRssKb = Number.NaN;
-    try {
-      const metrics = fs.readFileSync(metricsPath, "utf8").trim().split(/\s+/);
-      elapsedSeconds = Number.parseFloat(metrics[0]);
-      peakRssKb = Number.parseInt(metrics[1], 10);
-    } catch {}
-    try { fs.unlinkSync(metricsPath); } catch {}
-
-    const sample = {
-      iteration,
-      status: run.status,
-      signal: run.signal,
-      elapsedSeconds,
-      peakRssKb,
-    };
-    samples.push(sample);
-
-    if (run.error) failures.push(`${relative} iteration ${iteration}: ${run.error.message}`);
-    if (run.status !== 0) {
-      const stderr = String(run.stderr || "").slice(-4000);
-      const stdout = String(run.stdout || "").slice(-4000);
-      failures.push(`${relative} iteration ${iteration}: exit ${run.status}; stdout=${JSON.stringify(stdout)}; stderr=${JSON.stringify(stderr)}`);
+      );
+      let elapsedSeconds = Number.NaN;
+      let peakRssKb = Number.NaN;
+      try {
+        const metrics = fs.readFileSync(metricsPath, "utf8").trim().split(/\s+/u);
+        elapsedSeconds = Number.parseFloat(metrics[0]);
+        peakRssKb = Number.parseInt(metrics[1], 10);
+      } catch {}
+      try { fs.unlinkSync(metricsPath); } catch {}
+      samples.push({ iteration, status: run.status, signal: run.signal, elapsedSeconds, peakRssKb });
+      if (run.error) failures.push(`${relative} iteration ${iteration}: ${run.error.message}`);
+      if (run.status !== 0) {
+        const stderr = String(run.stderr || "").slice(-4000);
+        const stdout = String(run.stdout || "").slice(-4000);
+        failures.push(`${relative} iteration ${iteration}: exit ${run.status}; stdout=${JSON.stringify(stdout)}; stderr=${JSON.stringify(stderr)}`);
+      }
+      if (Number.isFinite(elapsedSeconds) && elapsedSeconds > perRunSecondsCeiling) failures.push(`${relative} iteration ${iteration}: ${elapsedSeconds}s exceeded ${perRunSecondsCeiling}s ceiling`);
+      if (Number.isFinite(peakRssKb) && peakRssKb > peakRssKbCeiling) failures.push(`${relative} iteration ${iteration}: ${peakRssKb} KiB exceeded ${peakRssKbCeiling} KiB RSS ceiling`);
     }
-    if (Number.isFinite(elapsedSeconds) && elapsedSeconds > perRunSecondsCeiling) {
-      failures.push(`${relative} iteration ${iteration}: ${elapsedSeconds}s exceeded ${perRunSecondsCeiling}s ceiling`);
-    }
-    if (Number.isFinite(peakRssKb) && peakRssKb > peakRssKbCeiling) {
-      failures.push(`${relative} iteration ${iteration}: ${peakRssKb} KiB exceeded ${peakRssKbCeiling} KiB RSS ceiling`);
-    }
+    const elapsed = samples.map(sample => sample.elapsedSeconds).filter(Number.isFinite);
+    const rss = samples.map(sample => sample.peakRssKb).filter(Number.isFinite);
+    results.push({
+      test: relative,
+      repeats,
+      passed: samples.every(sample => sample.status === 0),
+      medianElapsedSeconds: median(elapsed),
+      maximumElapsedSeconds: elapsed.length ? Math.max(...elapsed) : null,
+      medianPeakRssKb: median(rss),
+      maximumPeakRssKb: rss.length ? Math.max(...rss) : null,
+      samples,
+    });
   }
 
-  const elapsed = samples.map(sample => sample.elapsedSeconds).filter(Number.isFinite);
-  const rss = samples.map(sample => sample.peakRssKb).filter(Number.isFinite);
-  results.push({
-    test: relative,
-    repeats,
-    passed: samples.every(sample => sample.status === 0),
-    medianElapsedSeconds: median(elapsed),
-    maximumElapsedSeconds: elapsed.length ? Math.max(...elapsed) : null,
-    medianPeakRssKb: median(rss),
-    maximumPeakRssKb: rss.length ? Math.max(...rss) : null,
-    samples,
-  });
+  const totalSeconds = Number(process.hrtime.bigint() - started) / 1e9;
+  if (totalSeconds > totalSecondsCeiling) failures.push(`Complete runtime stress took ${totalSeconds.toFixed(2)}s, exceeding ${totalSecondsCeiling}s`);
+  const sourceText = fs.readFileSync(path.join(root, "src/MissionChief_Map_Command_Toolkit.user.js"), "utf8");
+  const report = {
+    schemaVersion: 2,
+    audit: "runtime-stress",
+    status: failures.length ? "failed" : "passed",
+    source: {
+      version: sourceText.match(/^\/\/\s*@version\s+([^\s]+)/mu)?.[1] || "unknown",
+      sha256: crypto.createHash("sha256").update(sourceText, "utf8").digest("hex"),
+    },
+    discovery: {
+      authority: ".github/scripts/run_userscript_preflight.sh",
+      authoritySha256: crypto.createHash("sha256").update(preflightSource, "utf8").digest("hex"),
+      canonicalRuntimeContracts: discoverRuntimeContracts(preflightSource),
+      explicitRuntimeContracts: tests.map(item => item[0]).filter(item => !discoverRuntimeContracts(preflightSource).includes(item)),
+      plannedContracts: tests.map(([test, repeats]) => ({ test, repeats })),
+    },
+    repeatedExecutions: results.reduce((sum, result) => sum + result.repeats, 0),
+    totalSeconds,
+    ceilings: { perRunSecondsCeiling, peakRssKbCeiling, totalSecondsCeiling },
+    failures,
+    results,
+  };
+  fs.writeFileSync(path.join(output, "runtime-stress.json"), `${JSON.stringify(report, null, 2)}\n`);
+  const markdown = [
+    "# Toolkit repeated runtime stress audit",
+    "",
+    `- **Status:** \`${report.status}\``,
+    `- **Toolkit version:** \`${report.source.version}\``,
+    `- **Source SHA-256:** \`${report.source.sha256}\``,
+    `- **Discovery authority:** \`${report.discovery.authority}\``,
+    `- **Discovered runtime contracts:** ${report.discovery.canonicalRuntimeContracts.length}`,
+    `- **Explicit heavier contracts:** ${report.discovery.explicitRuntimeContracts.length}`,
+    `- **Repeated executions:** ${report.repeatedExecutions}`,
+    `- **Total elapsed:** ${totalSeconds.toFixed(2)} seconds`,
+    `- **Per-run ceiling:** ${perRunSecondsCeiling} seconds`,
+    `- **Peak RSS ceiling:** ${peakRssKbCeiling.toLocaleString("en-GB")} KiB`,
+    "",
+    "| Runtime contract | Repeats | Median | Maximum | Median RSS | Maximum RSS |",
+    "|---|---:|---:|---:|---:|---:|",
+    ...results.map(result => `| \`${result.test}\` | ${result.repeats} | ${result.medianElapsedSeconds.toFixed(2)}s | ${result.maximumElapsedSeconds?.toFixed(2) ?? "—"}s | ${Math.round(result.medianPeakRssKb || 0).toLocaleString("en-GB")} KiB | ${Math.round(result.maximumPeakRssKb || 0).toLocaleString("en-GB")} KiB |`),
+    "",
+    "## Failures",
+    "",
+    ...(failures.length ? failures.map(failure => `- ${failure}`) : ["- None."]),
+    "",
+  ];
+  fs.writeFileSync(path.join(output, "runtime-stress.md"), markdown.join("\n"));
+  console.log(JSON.stringify(report, null, 2));
+  process.exitCode = failures.length ? 1 : 0;
+  return report;
 }
 
-const totalSeconds = Number(process.hrtime.bigint() - started) / 1e9;
-if (totalSeconds > totalSecondsCeiling) {
-  failures.push(`Complete runtime stress took ${totalSeconds.toFixed(2)}s, exceeding ${totalSecondsCeiling}s`);
-}
-
-const report = {
-  schemaVersion: 1,
-  audit: "runtime-stress",
-  status: failures.length ? "failed" : "passed",
-  repeatedExecutions: results.reduce((sum, result) => sum + result.repeats, 0),
-  totalSeconds,
-  ceilings: { perRunSecondsCeiling, peakRssKbCeiling, totalSecondsCeiling },
-  failures,
-  results,
-};
-fs.writeFileSync(path.join(output, "runtime-stress.json"), `${JSON.stringify(report, null, 2)}\n`);
-
-const markdown = [
-  "# Toolkit repeated runtime stress audit",
-  "",
-  `- **Status:** \`${report.status}\``,
-  `- **Repeated executions:** ${report.repeatedExecutions}`,
-  `- **Total elapsed:** ${totalSeconds.toFixed(2)} seconds`,
-  `- **Per-run ceiling:** ${perRunSecondsCeiling} seconds`,
-  `- **Peak RSS ceiling:** ${peakRssKbCeiling.toLocaleString("en-GB")} KiB`,
-  "",
-  "| Runtime contract | Repeats | Median | Maximum | Median RSS | Maximum RSS |",
-  "|---|---:|---:|---:|---:|---:|",
-  ...results.map(result => `| \`${result.test}\` | ${result.repeats} | ${result.medianElapsedSeconds.toFixed(2)}s | ${result.maximumElapsedSeconds?.toFixed(2) ?? "—"}s | ${Math.round(result.medianPeakRssKb || 0).toLocaleString("en-GB")} KiB | ${Math.round(result.maximumPeakRssKb || 0).toLocaleString("en-GB")} KiB |`),
-  "",
-  "## Failures",
-  "",
-  ...(failures.length ? failures.map(failure => `- ${failure}`) : ["- None."]),
-  "",
-];
-fs.writeFileSync(path.join(output, "runtime-stress.md"), markdown.join("\n"));
-console.log(JSON.stringify(report, null, 2));
-process.exitCode = failures.length ? 1 : 0;
+const isMain = process.argv[1] && path.resolve(process.argv[1]) === fileURLToPath(import.meta.url);
+if (isMain) runRuntimeStressAudit();
