+++ .github/scripts/run_userscript_preflight.sh
@@ -8,6 +8,7 @@ for contract in .github/scripts/test_financial_ledger_contract.py .github/script
 python3 .github/scripts/test_release_pipeline_v4.py
 python3 .github/scripts/test_consolidated_pr_gate.py
 python3 .github/scripts/test_path_aware_blocking.py
+python3 .github/scripts/test_issue588_v831_performance_baseline.py
 node .github/scripts/test_transport_sweep_runtime.js
 node .github/scripts/test_issue523_transport_sweep_progress_runtime.js
 node .github/scripts/test_issue527_transport_sweep_skipped_patients_runtime.js
