+++ docs/audits/issue-588/performance-budget-report.json
@@ -1,0 +1,41 @@
+{
+  "schemaVersion": 2,
+  "policyRevision": "2026-07-29-v831-exact-artifact-production-proof",
+  "candidate": {
+    "bytes": 1654208,
+    "lines": 25228,
+    "nonempty_lines": 24072,
+    "css_bytes": 704606,
+    "css_rule_blocks": 5464,
+    "mutation_observer_constructions": 11,
+    "resize_observer_constructions": 4,
+    "broad_subtree_observers": 9,
+    "document_wide_subtree_observers": 3,
+    "network_request_calls": 5,
+    "set_interval_calls": 1,
+    "set_timeout_calls": 3,
+    "mutation_observers": 8,
+    "resize_observers": 1,
+    "request_animation_frame_calls": 1,
+    "event_listener_calls": 44,
+    "query_selector_calls": 228,
+    "startup_hook_calls": 5,
+    "get_element_by_id_calls": 78,
+    "inner_html_writes": 16,
+    "runtime_interval_calls": 1,
+    "runtime_timeout_calls": 81,
+    "runtime_animation_frame_calls": 10,
+    "runtime_listener_calls": 26,
+    "runtime_observer_track_calls": 13,
+    "runtime_idle_calls": 3
+  },
+  "base": null,
+  "findings": [
+    {
+      "level": "warning",
+      "metric": "resize_observer_constructions",
+      "message": "ResizeObserver constructions including aliases uses 4 of 4 (100.0%); review threshold is 100.0%."
+    }
+  ],
+  "result": "success"
+}
